split calculation
